requirements.txt
asgiref==3.8.1
astroid==3.2.2
attrs==23.2.0
black==24.4.2
click==8.1.7
colorama==0.4.6
coverage==7.5.3
dill==0.3.8
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
exceptiongroup==1.2.1
Faker==25.8.0
flake8==7.0.0
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.5.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.8.0
pylint==3.2.3
pytest==8.2.2
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
tomlkit==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
